+++ src/scenes/BootScene.js
@@ -124,9 +124,11 @@ export default class BootScene extends Phaser.Scene {
 
     // Stance icons
     this.load.image('icon_tempest',  'Images/TempestIcon2.png');
-    this.load.image('icon_ruin',     'Images/RuinIcon.png');
-    this.load.image('icon_fortress', 'Images/FortressIcon.png');
+    this.load.image('icon_ruin',     'Images/RuinIcon2.png');
+    this.load.image('icon_fortress', 'Images/Fortressstance2.png');
     this.load.image('icon_flurry_button', 'Images/FlurryIcon2.png');
+    this.load.image('icon_smash_button', 'Images/SmashIcon_.png');
+    this.load.image('icon_bulwark_button', 'Images/BulwarkIcon.png');
     this.load.image('player001_default',        'Images/Player Images/armor001/Player001_default.png');
     this.load.image('player001_strongpunch',    'Images/Player Images/armor001/Player001_strongpunch.png');
     this.load.image('player001_jumpkick',       'Images/Player Images/armor001/Player001_jumpkick.png');
@@ -284,7 +286,7 @@ export default class BootScene extends Phaser.Scene {
       { keys: ['goblinwarrior001_default', 'goblinwarrior001_reaction', 'goblinwarrior001_attack', 'goblinwarrior001_dead'], w: 320, h: 384 },
       { keys: ['crackedarmor001'], w: 512, h: 512 },
       { keys: ['icon_tempest', 'icon_ruin', 'icon_fortress'], w: 128, h: 128 },
-      { keys: ['icon_flurry_button'], w: 64, h: 64 },
+      { keys: ['icon_flurry_button', 'icon_smash_button', 'icon_bulwark_button'], w: 128, h: 128 },
     ];
     for (const { keys, w, h } of enemyGroups) {
       for (const key of keys) {